### 2021-11-08 #483 @Yangcq-Ivy
NicheBarcoding: Species Identification using DNA Barcodes Integrated with Environmental Niche Models.
**Reviewers:**
Submitting Author Name: Cai-qing Yang
Submitting Author Github Handle: @Yangcq-Ivy
Other Package Authors Github handles: @zhangab2020
Repository:  https://github.com/Yangcq-Ivy/NicheBarcoding
Version submitted: 1.0
Submission type: Standard

Version accepted: TBD
---

```
Package: NicheBarcoding
Title: Niche-model-Based Species Identification
Version: 1.0
    c(person(given = "Cai-qing",
             family = "YANG",
             email = "yangcq_ivy@163.com"),
      person(given = "Xin-hai",
             family = "Li",
             role = "aut",
             email = "lixh@ioz.ac.cn"),
      person(given = "Michael christopher",
             family = "ORR",
             role = "aut",
             email = "michael.christopher.orr@gmail.com"),
      person(given = "Ai-bing",
             family = "ZHANG",
             role = "aut",
             email = "Zhangab2008@cnu.edu.cn"))
Description: Species Identification using DNA Barcodes Integrated with
             Environmental Niche Models.
Language: es
LazyDataCompression: bzip2
Imports: ape, dismo, e1071(>= 1.7-7), maps, pROC, randomForest, raster,
        rJava, spider, vegan
Author: Cai-qing YANG [aut, cre],
  Xin-hai Li [aut],
  Michael christopher ORR [aut],
  Ai-bing ZHANG [aut]
Maintainer: Cai-qing YANG <yangcq_ivy@163.com>
NeedsCompilation: no
Packaged: 2021-11-05 10:00:06 UTC; zhang
```

NicheBarcoding provides a practical and user-friendly workflow that fitting the theory of integrated taxonomy for global biodiversity research.
Here, the global bio-climatic variables are used to calibrate the potential mis-identification from DNA barcoding.

NicheBarcoding can provide a quick, comprehensive, and accurate species identification for global biodiversity researchers in their ecological and evolutionary studies.

No.
Current packages can only achieve species identification of DNA barcoding alone, but it does not solve its inherent shortcomings and lacks evidence from the ecological perspective.
Although some packages can be used to build niche models, few are used for the problem of multi-species identification.
NicheBarcoding not only integrates both two aspects, but also provides algorithm solutions for coping different scenarios.
The affiliating of ecological niche model significantly improves the identification success rate compares to the sole use of DNA barcoding, which makes up the intrinsic defects of both two sides.

Yes.
