### 2021-11-08 #483 @ldecicco-USGS
Hi @Yangcq-Ivy ! Thanks for the submission. We need the submission to be a git repository of the package... not a repository with a single  binary tarball of the package. For example, notice the structure of one of the most recent packages in rOpenSci is shown here:
https://github.com/ropensci/excluder

Let me know when you restructure your repository, and I'll kick off the initial checks again, thanks!
