### 2021-11-10 #483 @Yangcq-Ivy
Dear Laura,

Sorry for the trouble! I have just re-uploaded the package into a git repository.
Please check them through the previous link: https://github.com/Yangcq-Ivy/NicheBarcoding
Let me know if there is anything else I can do! Thanks a lot!

Yours,
Cai-qing Yang

At 2021-11-09 02:54:18, "Laura DeCicco" ***@***.***> wrote:

Hi @Yangcq-Ivy ! Thanks for the submission. We need the submission to be a git repository of the package... not a repository with a single binary tarball of the package. For example, notice the structure of one of the most recent packages in rOpenSci is shown here:
https://github.com/ropensci/excluder

Let me know when you restructure your repository, and I'll kick off the initial checks again, thanks!
