### 2021-11-12 #483 @Yangcq-Ivy
Dear reviewer,

We have updated / supplemented the relevant documents according to the check list.
Please find them through the link: https://github.com/Yangcq-Ivy/NicheBarcoding
Thanks so much for the checks on the package!

Sincerely,
Cai-qing Yang

At 2021-11-10 21:15:59, "ropensci-review-bot" ***@***.***> wrote:

Checks for NicheBarcoding (v1.0)

git hash: 9abfa9fa

✔️ Package name is available
✖️ does not have a 'CITATION' file.
✖️ does not have a 'codemeta.json' file.
✖️ does not have a 'contributing' file.
✔️ uses 'roxygen2'.
✖️ 'DESCRIPTION' does not have a URL field.
✖️ 'DESCRIPTION' does not have a BugReports field.
✖️ Package has no HTML vignettes
✔️ All functions have examples.
✖️ Continuous integration checks unavailable (no URL in 'DESCRIPTION').
✖️ Package coverage is 0% (should be at least 75%).
✔️ R CMD check found no errors.
✔️ R CMD check found no warnings.

Important: All failing checks above must be addressed prior to proceeding

Package License: GPL (>= 3)

1. Statistical Properties

Details of statistical properties (click to open)

The package has:

code in R (100% in 10 files) and
4 authors
no vignette
3 internal data files
10 imported packages
9 exported functions (median 82 lines of code)
16 non-exported functions in R (median 53 lines of code)

The following terminology is used:

loc = "Lines of Code"
fn = "function"
exp/not_exp = exported / not exported

The final measure (fn_call_network_size) is the total number of calls between functions (in R), or more abstract relationships between code objects in other languages. Values are flagged as "noteworthy" when they lie in the upper or lower 5th percentile.

| measure | value | percentile | noteworthy |
| files_R | 10 | 55.4 | |
| files_vignettes | 0 | 0.0 | TRUE |
| files_tests | 0 | 0.0 | TRUE |
| loc_R | 961 | 65.1 | |
| num_vignettes | 0 | 0.0 | TRUE |
| data_size_total | 25844602 | 100.0 | TRUE |
| data_size_median | 37680 | 87.0 | |
| n_fns_r | 25 | 26.1 | |
| n_fns_r_exported | 9 | 38.9 | |
| n_fns_r_not_exported | 16 | 24.5 | |
| n_fns_per_file_r | 1 | 18.0 | |
| num_params_per_fn | 6 | 79.2 | |
| loc_per_fn_r | 74 | 97.6 | TRUE |
| loc_per_fn_r_exp | 82 | 90.3 | |
| loc_per_fn_r_not_exp | 53 | 95.9 | TRUE |
| rel_whitespace_R | 23 | 71.4 | |
| doclines_per_fn_exp | 54 | 67.3 | |
| doclines_per_fn_not_exp | 0 | 0.0 | TRUE |
| fn_call_network_size | 20 | 39.5 | |

1a. Network visualisation

Click to see the interactive network visualisation of calls between objects in package

2. goodpractice and other checks
Details of goodpractice and other checks (click to open)

3b. goodpractice results
R CMD check with rcmdcheck

R CMD check generated the following notes:

checking installed package size ... NOTE
installed size is 24.8Mb
sub-directories of 1Mb or more:
data 24.7Mb
checking dependencies in R code ... NOTE
Namespace in Imports field not imported from: ‘rJava’
All declared Imports should be used.

cyclocomp
description_url
description_bugreports
rcmdcheck_imports_not_imported_from
rcmdcheck_reasonable_installed_size
Test coverage with covr

Package coverage: 0

| file | coverage |
| R/extractSpeInfo.R | 0% |
| R/monophyly.prop.R | 0% |
| R/NBSI.R | 0% |
| R/NBSI2.R | 0% |
| R/niche.Model.Build.R | 0% |
| R/niche.PCA.R | 0% |
| R/pseudo.absent.points.R | 0% |
| R/pseudo.present.points.R | 0% |
| R/spe.mantel.test.R | 0% |
Cyclocomplexity with cyclocomp

| function | cyclocomplexity |
| NBSI2 | 52 |
| NBSI | 51 |
| niche.Model.Build | 29 |
Static code analyses with lintr

lintr found no issues with this package!

Package Versions

| package | version |
| pkgstats | 0.0.2.72 |
| pkgcheck | 0.0.2.107 |

Editor-in-Chief Instructions:

Processing may not proceed until the items marked with ✖️ have been resolved.
