### 2021-11-17 #483 @ldecicco-USGS
Hi @Yangcq-Ivy and @zhangab2020. Before we can make a good choice on if the `NicheBarcoding` package is a good fit for rOpenSci, you'll need to add some more information in a vignette, as well as adding tests via the `testthat` package. See:
https://devguide.ropensci.org/ci.html for a lot more information, as well as a good overview of `testthat` here:
https://testthat.r-lib.org/

In the meantime, I'll look for editors who might be familiar with DNA barcoding processes.
