+++ cran-comments.md
@@ -1,0 +1,20 @@
+## Test environments
+* centOS6.4, 4.1.1 (2021-08-10)
+* using platform: x86_64-w64-mingw32 (64-bit)
+* using session charset: CP936
+
+## R CMD check results
+There were no ERRORs,or WARNING.
+There was 1 NOTE:
+
+checking installed package size ... NOTE
+    installed size is 24.9Mb
+    sub-directories of 1Mb or more:
+      data  24.7Mb
+
+* checking CRAN incoming feasibility ... NOTE
+Maintainer: 'Cai-qing YANG <yangcq_ivy@163.com>'
+New submission
+
+## Downstream dependencies
+No downtream dependencies.
