+++ DESCRIPTION
@@ -1,0 +1,44 @@
+Package: NicheBarcoding
+Title: Niche-model-Based Species Identification
+Version: 1.0
+Authors@R: 
+    c(person(given = "Cai-qing",
+             family = "YANG",
+             role = c("aut", "cre"),
+             email = "yangcq_ivy@163.com"),
+      person(given = "Xin-hai",
+             family = "Li",
+             role = "aut",
+             email = "lixh@ioz.ac.cn"),
+      person(given = "Michael christopher",
+             family = "ORR",
+             role = "aut",
+             email = "michael.christopher.orr@gmail.com"),
+      person(given = "Ai-bing",
+             family = "ZHANG",
+             role = "aut",
+             email = "Zhangab2008@cnu.edu.cn"))
+Description: Species Identification using DNA Barcodes Integrated with 
+             Environmental Niche Models.
+License: GPL (>= 3)
+Encoding: UTF-8
+Language: es
+LazyData: true
+LazyDataCompression: bzip2
+Roxygen: list(markdown = TRUE)
+RoxygenNote: 7.1.2
+Imports: ape, dismo, e1071 (>= 1.7-7), maps, pROC, randomForest,
+        raster, rJava, spider, vegan
+Depends: R (>= 2.10)
+VignetteBuilder: knitr
+Author: Cai-qing YANG [aut, cre],
+  Xin-hai Li [aut],
+  Michael christopher ORR [aut],
+  Ai-bing ZHANG [aut]
+Maintainer: Cai-qing YANG <yangcq_ivy@163.com>
+URL: https://github.com/Yangcq-Ivy/NicheBarcoding
+BugReports: https://github.com/Yangcq-Ivy/NicheBarcoding
+NeedsCompilation: no
+Packaged: 2021-12-15 09:14:59 UTC; zhang
+Suggests: rmarkdown, covr, testthat (>= 3.0.0), knitr
+Config/testthat/edition: 3
