### 2021-12-17 #483 @Yangcq-Ivy
Dear reviewers,
We have just re-uploaded our package with improving the test coverage and adding the .yaml file for the continuous integration.
Thanks very much for your kindly check! Looking forward to more suggestions for improving the package.
