### 2022-01-14 #483 @ldecicco-USGS
Hi @Yangcq-Ivy , sorry for the delay, this issues slipped through my radar. I'll re-run the checks. I'd still really like to read more in the vignettes before deciding if the package fits the rOpenSci scope (and therefore assigning a reviewer).
