### 2022-01-28 #483 @ldecicco-USGS
I will check on what is causing the error "HTML variable [author1] is missing.

I can see you have made some significant improvements to your package since submission. The initial submission was difficult for me to judge the "fit" to rOpenSci. I am going to close this issue and classify the package as not currently fitting the rOpenSci . The definitions can be found here:
https://devguide.ropensci.org/policies.html#package-categories

Here is the feedback I received when consulting other editors with a broader knowledge of DNA barcoding:
The authors have checked off workflow automation, scientific software wrappers and geospatial data as their categories for “Scope”. However I do not see how they fulfil the fit as we have defined in the dev guide.

For workflow automation: “Tools that automate and link together workflows” - The authors of NicheBarcoding are providing tools that run a statistical model to estimate the probability of an identified organism being in the defined region. This doesn’t seem like a workflow automation tool.

For scientific software wrappers: “Packages that wrap non-R utility programs used for scientific research” I don’t see the use of any non-R programs here (but maybe I missed it)

For geospatial data: “We accept packages focused on accessing geospatial data, manipulating geospatial data, and converting between geospatial data formats.” - True they are using data retrieved using {raster} and {terra} but the package’s main focus is not accessing, manipulating or converting geospatial data, but accessing it as needed for use in their model.

If you would like to re-submit, especially if you adjust the package to make it useful to a wider audience, be sure to clarify how you see it fitting within the rOpenSci categories.

Thanks again for the submission!
